Add initial Jakarta EE structure and classes
This commit includes the initial Jakarta EE structure, resources, and classes. Dockerfile is also added for containerization. Beans.xml, index.html, and payara logo are added in webapp directory. A Hello class is generated along with HelloApplication and HelloWorldResource for RESTful service. The .gitignore is also updated for Maven and IDEA.